Lint test_support with Whitaker and narrow its exemption (#462)
`test_support` is excluded from the root Cargo workspace, so a
workspace-root `cargo dylint` never reached it and its entry in the root
`dylint.toml` was inert. Run the suite a second time from `test_support/`
in `make lint-whitaker` so the crate is genuinely gated, and give it a
`test_support/dylint.toml` whose only `no_std_fs_operations` exemption is
the `test_support::fs` boundary module.

Clearing the resulting backlog:

- Route `check_ninja`, `exec`, `manifest`, and the crate root through
  `crate::fs` instead of calling `std::fs` directly, so `test_support::fs`
  is the crate's single ambient boundary.
- Make the `original_dir` rstest fixture fallible rather than expecting
  inside it, and recover the localiser test lock from poisoning the way
  `env_lock` already does. Both sites are fixtures, which the lint cannot
  recognize as test code.
- Replace the `assert_underlying_lock_is_*` helper functions with macros
  so failures report the calling test's line number.
- Extract `http`'s test module to `http_tests.rs`, bringing the module
  back under `module_max_lines`.
- Add the missing inner doc comments and drop an unused `anyhow` import.
